Added sprite and organized
