+++ Assets/Scripts/Background/ParallaxBackground.cs.meta
