+++ Assets/Scripts/Background/ParallaxCamera.cs
