+++ Assets/Scripts/Background/ParallaxLayer.cs
