+++ Assets/Scripts/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8fd15f88e60f3ab419e1850b7a4bfb9b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
