+++ Assets/Scripts/Menus/QuitMenu.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class QuitMenu : MonoBehaviour
+{
+    // Start is called before the first frame update
+    public void Quit()
+    {
+        Application.Quit();
+        Debug.Log("Quit Game");
+    }
+}
