--- Assets/Scripts/QuitMenu.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class QuitMenu : MonoBehaviour
-{
-    // Start is called before the first frame update
-    public void Quit()
-    {
-        Application.Quit();
-        Debug.Log("Quit Game");
-    }
-}
